requirements.txt
click>=8.1.3
pydantic>=1.10.7
jsonschema>=4.17.3
datamodel-code-generator>=0.16.0
jinja2>=3.1.2
openai>=0.27.8
langchain>=0.0.200
pinecone-client>=2.2.1
mcp-use>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
pytest>=7.3.1
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
flake8>=6.0.0
black>=23.9.1
isort>=5.12.0
pre-commit>=3.4.0